@@ -24,10 +24,6 @@ module AuthorizeNet::API
   class ArrayOfBatchStatisticType < ::Array
   end
   
-  # {AnetApi/xml/v1/schema/AnetApiSchema.xsd}ArrayOfSetting
-  class ArrayOfSetting < ::Array
-  end
-  
   # {AnetApi/xml/v1/schema/AnetApiSchema.xsd}emailSettingsType
   #   setting - SettingType
   #   xmlattr_version - SOAP::SOAPInteger
@@ -244,10 +240,14 @@ module AuthorizeNet::API
     include ROXML
     xml_accessor :marketType
     xml_accessor :deviceType
+    xml_accessor :customerSignature
+    xml_accessor :terminalNumber
   
-    def initialize(marketType = nil, deviceType = nil)
+    def initialize(marketType = nil, deviceType = nil, customerSignature = nil, terminalNumber = nil)
       @marketType = marketType
       @deviceType = deviceType
+      @customerSignature = customerSignature
+      @terminalNumber = terminalNumber
     end
   end
   
@@ -313,12 +313,14 @@ module AuthorizeNet::API
     xml_accessor :expirationDate
     xml_accessor :cardType
     xml_accessor :cardArt
+    xml_accessor :issuerNumber
   
-    def initialize(cardNumber = nil, expirationDate = nil, cardType = nil)
+    def initialize(cardNumber = nil, expirationDate = nil, cardType = nil, cardArt = nil, issuerNumber = nil)
       @cardNumber = cardNumber
       @expirationDate = expirationDate
       @cardType = cardType
       @cardArt = cardArt
+      @issuerNumber = issuerNumber
     end
   end
   
@@ -784,6 +786,7 @@ end
   #   impersonationAuthentication - ImpersonationAuthenticationType
   #   fingerPrint - FingerPrintType
   #   mobileDeviceId - SOAP::SOAPString
+  #   accessToken - SOAP::SOAPString
   class MerchantAuthenticationType
     include ROXML
     xml_accessor :name
@@ -793,8 +796,9 @@ end
     xml_accessor :impersonationAuthentication, :as => ImpersonationAuthenticationType
     xml_accessor :fingerPrint, :as => FingerPrintType
     xml_accessor :mobileDeviceId
+	xml_accessor :accessToken
   
-    def initialize(name = nil, transactionKey = nil, sessionToken = nil, password = nil, impersonationAuthentication = nil, fingerPrint = nil, mobileDeviceId = nil)
+    def initialize(name = nil, transactionKey = nil, sessionToken = nil, password = nil, impersonationAuthentication = nil, fingerPrint = nil, mobileDeviceId = nil, accessToken = nil)
       @name = name
       @transactionKey = transactionKey
       @sessionToken = sessionToken
@@ -802,6 +806,7 @@ end
       @impersonationAuthentication = impersonationAuthentication
       @fingerPrint = fingerPrint
       @mobileDeviceId = mobileDeviceId
+      @accessToken = accessToken
     end
   end
   
@@ -992,7 +997,41 @@ end
     end
   end
   
-     
+    # {AnetApi/xml/v1/schema/AnetApiSchema.xsd}subMerchantType
+  #   identifier - SOAP::SOAPString
+  #   doingBusinessAs - SOAP::SOAPString
+  #   paymentServiceProviderName - SOAP::SOAPString
+  #   paymentServiceFacilitator - SOAP::SOAPString
+  #   streetAddress - SOAP::SOAPString
+  class SubMerchantType
+    include ROXML
+    xml_accessor :identifier
+    xml_accessor :doingBusinessAs
+    xml_accessor :paymentServiceProviderName
+    xml_accessor :paymentServiceFacilitator
+    xml_accessor :streetAddress
+    xml_accessor :phone
+    xml_accessor :email
+    xml_accessor :postalCode
+    xml_accessor :city
+    xml_accessor :regionCode
+    xml_accessor :countryCode
+  
+    def initialize(identifier = nil, doingBusinessAs = nil, paymentServiceProviderName = nil, paymentServiceFacilitator = nil, streetAddress = nil, phone = nil, email = nil, postalCode = nil, city = nil, regionCode = nil, countryCode = nil )
+      @identifier = identifier
+      @doingBusinessAs = doingBusinessAs
+      @paymentServiceProviderName = paymentServiceProviderName
+      @paymentServiceFacilitator = paymentServiceFacilitator
+      @streetAddress = streetAddress
+      @phone = phone
+      @email = email
+      @postalCode = postalCode
+      @city = city
+      @regionCode = regionCode
+      @countryCode = countryCode
+    end
+  end
+  
   # {AnetApi/xml/v1/schema/AnetApiSchema.xsd}settingType
   #   settingName - SOAP::SOAPString
   #   settingValue - SOAP::SOAPString
@@ -1014,7 +1053,7 @@ end
       @settings = settings
     end
   end
-  
+
   # {AnetApi/xml/v1/schema/AnetApiSchema.xsd}permissionType
   #   permissionName - SOAP::SOAPString
   class PermissionType
@@ -1678,6 +1717,8 @@ end
   #   mobileDeviceId - SOAP::SOAPString
   #   subscription - SubscriptionPaymentType
   #   hasReturnedItems - SOAP::SOAPBoolean
+  #   fraudInformation - FraudInformationType
+  #   profile - CustomerProfileIdType
   class TransactionSummaryType
     include ROXML
     xml_accessor :transId
@@ -1693,11 +1734,12 @@ end
     xml_accessor :marketType
     xml_accessor :product
     xml_accessor :mobileDeviceId
-    xml_accessor :subscription
+    xml_accessor :subscription, :as => SubscriptionPaymentType
     xml_accessor :hasReturnedItems
     xml_accessor :fraudInformation
+	xml_accessor :profile, :as => CustomerProfileIdType
   
-    def initialize(transId = nil, submitTimeUTC = nil, submitTimeLocal = nil, transactionStatus = nil, invoiceNumber = nil, firstName = nil, lastName = nil, accountType = nil, accountNumber = nil, settleAmount = nil, marketType = nil, product = nil, mobileDeviceId = nil, subscription = nil, hasReturnedItems = nil, fraudInformation = nil)
+    def initialize(transId = nil, submitTimeUTC = nil, submitTimeLocal = nil, transactionStatus = nil, invoiceNumber = nil, firstName = nil, lastName = nil, accountType = nil, accountNumber = nil, settleAmount = nil, marketType = nil, product = nil, mobileDeviceId = nil, subscription = nil, hasReturnedItems = nil, fraudInformation = nil, profile = nil)
       @transId = transId
       @submitTimeUTC = submitTimeUTC
       @submitTimeLocal = submitTimeLocal
@@ -1714,6 +1756,27 @@ end
       @subscription = subscription
       @hasReturnedItems = hasReturnedItems
       @fraudInformation = fraudInformation
+	  @profile = profile
+    end
+  end
+
+  # {AnetApi/xml/v1/schema/AnetApiSchema.xsd}arbTransaction
+  #   transId - (any)
+  #   response - SOAP::SOAPString
+  #   submitTimeUTC - SOAP::SOAPDateTime
+  #   payNum - SOAP::SOAPString
+  class ArbTransaction
+    include ROXML
+    xml_accessor :transId
+    xml_accessor :response
+    xml_accessor :submitTimeUTC
+    xml_accessor :payNum
+  
+    def initialize(transId = nil, response = nil, submitTimeUTC = nil, payNum = nil)
+      @transId = transId
+      @response = response
+      @submitTimeUTC = submitTimeUTC
+      @payNum = payNum
     end
   end
 
@@ -1726,7 +1789,16 @@ end
      @transaction = transaction
     end
   end
-  
+ 
+   # {AnetApi/xml/v1/schema/AnetApiSchema.xsd}ARBTransactionList
+   class ARBTransactionList < ::Array
+    include ROXML
+    xml_accessor :arbTransaction, :as => [ArbTransaction]
+
+    def initialize(arbTransaction = [])
+     @arbTransaction = arbTransaction
+    end
+  end
 
   # {AnetApi/xml/v1/schema/AnetApiSchema.xsd}batchStatisticType
   #   accountType - SOAP::SOAPString
@@ -1926,6 +1998,7 @@ end
   #   returnedItems - ArrayOfReturnedItem
   #   solution - SolutionType
   #   emvDetails - TransactionDetailsType::EmvDetails
+  #   profile - CustomerProfileIdType
   class TransactionDetailsType
     include ROXML
     # inner class for member: EmvDetails
@@ -1959,7 +2032,7 @@ end
     xml_accessor :transactionStatus
     xml_accessor :responseCode
     xml_accessor :responseReasonCode
-    xml_accessor :subscription
+    xml_accessor :subscription, :as => SubscriptionPaymentType
     xml_accessor :responseReasonDescription
     xml_accessor :authCode
     xml_accessor :aVSResponse
@@ -1990,8 +2063,12 @@ end
     xml_accessor :returnedItems, :as => ArrayOfReturnedItem
     xml_accessor :solution
     xml_accessor :emvDetails, :as => EmvDetails
+  xml_accessor :profile, :as => CustomerProfileIdType
+  xml_accessor :surcharge, :as => ExtendedAmountType
+  xml_accessor :employeeId
+  xml_accessor :tip, :as => ExtendedAmountType
   
-    def initialize(transId = nil, refTransId = nil, splitTenderId = nil, submitTimeUTC = nil, submitTimeLocal = nil, transactionType = nil, transactionStatus = nil, responseCode = nil, responseReasonCode = nil, subscription = nil, responseReasonDescription = nil, authCode = nil, aVSResponse = nil, cardCodeResponse = nil, cAVVResponse = nil, fDSFilterAction = nil, fDSFilters = nil, batch = nil, order = nil, requestedAmount = nil, authAmount = nil, settleAmount = nil, tax = nil, shipping = nil, duty = nil, lineItems = nil, prepaidBalanceRemaining = nil, taxExempt = nil, payment = nil, customer = nil, billTo = nil, shipTo = nil, recurringBilling = nil, customerIP = nil, product = nil, marketType = nil, mobileDeviceId = nil, returnedItems = nil, solution = nil, emvDetails = nil)
+    def initialize(transId = nil, refTransId = nil, splitTenderId = nil, submitTimeUTC = nil, submitTimeLocal = nil, transactionType = nil, transactionStatus = nil, responseCode = nil, responseReasonCode = nil, subscription = nil, responseReasonDescription = nil, authCode = nil, aVSResponse = nil, cardCodeResponse = nil, cAVVResponse = nil, fDSFilterAction = nil, fDSFilters = nil, batch = nil, order = nil, requestedAmount = nil, authAmount = nil, settleAmount = nil, tax = nil, shipping = nil, duty = nil, lineItems = nil, prepaidBalanceRemaining = nil, taxExempt = nil, payment = nil, customer = nil, billTo = nil, shipTo = nil, recurringBilling = nil, customerIP = nil, product = nil, marketType = nil, mobileDeviceId = nil, returnedItems = nil, solution = nil, emvDetails = nil, profile = nil, surcharge = nil, employeeId = nil, tip = nil)
       @transId = transId
       @refTransId = refTransId
       @splitTenderId = splitTenderId
@@ -2032,6 +2109,10 @@ end
       @returnedItems = returnedItems
       @solution = solution
       @emvDetails = emvDetails
+    @profile = profile
+    @surcharge = surcharge
+    @employeeId = employeeId
+    @tip = tip
     end
   end
   
@@ -2090,6 +2171,7 @@ end
   #   secureAcceptance - TransactionResponse::secureAcceptance
   #   emvResponse - TransactionResponse::emvResponse
   #   transHashSha2 - SOAP::SOAPString
+  #   profile - CustomerProfileIdType
   class TransactionResponse
     include ROXML
     # inner class for member: prePaidCard
@@ -2269,8 +2351,9 @@ end
     xml_accessor :secureAcceptance, :as => SecureAcceptance
     xml_accessor :emvResponse, :as => EmvResponse
     xml_accessor :transHashSha2
+	xml_accessor :profile, :as => CustomerProfileIdType
   
-    def initialize(responseCode = nil, rawResponseCode = nil, authCode = nil, avsResultCode = nil, cvvResultCode = nil, cavvResultCode = nil, transId = nil, refTransID = nil, transHash = nil, testRequest = nil, accountNumber = nil, accountType = nil, splitTenderId = nil, prePaidCard = nil, messages = nil, errors = nil, splitTenderPayments = nil, userFields = nil, shipTo = nil, secureAcceptance = nil, emvResponse = nil, transHashSha2 = nil)
+    def initialize(responseCode = nil, rawResponseCode = nil, authCode = nil, avsResultCode = nil, cvvResultCode = nil, cavvResultCode = nil, transId = nil, refTransID = nil, transHash = nil, testRequest = nil, accountNumber = nil, accountType = nil, splitTenderId = nil, prePaidCard = nil, messages = nil, errors = nil, splitTenderPayments = nil, userFields = nil, shipTo = nil, secureAcceptance = nil, emvResponse = nil, transHashSha2 = nil, profile = nil)
       @responseCode = responseCode
       @rawResponseCode = rawResponseCode
       @authCode = authCode
@@ -2293,6 +2376,7 @@ end
       @secureAcceptance = secureAcceptance
       @emvResponse = emvResponse
       @transHashSha2 = transHashSha2
+	  @profile = profile
     end
   end
   
@@ -2522,7 +2606,7 @@ end
   #   cardCode - (any)
   class PaymentProfile
     include ROXML
-    xml_accessor :paymentProfileId, :as => Fixnum
+    xml_accessor :paymentProfileId, :as => Integer
     xml_accessor :cardCode
   
     def initialize(paymentProfileId = nil, cardCode = nil)
@@ -2539,9 +2623,9 @@ end
   class CustomerProfilePaymentType
     include ROXML
     xml_accessor :createProfile
-    xml_accessor :customerProfileId, :as => Fixnum
+    xml_accessor :customerProfileId, :as => Integer
     xml_accessor :paymentProfile, :as => PaymentProfile
-    xml_accessor :shippingProfileId, :as => Fixnum
+    xml_accessor :shippingProfileId, :as => Integer
   
     def initialize(createProfile = nil, customerProfileId = nil, paymentProfile = nil, shippingProfileId = nil)
       @createProfile = createProfile
@@ -2734,6 +2818,8 @@ end
     MerchantEmail = SettingNameEnum.new("merchantEmail")
     RecurringBilling = SettingNameEnum.new("recurringBilling")
     TestRequest = SettingNameEnum.new("testRequest")
+    TypeEmailReceipt = SettingNameEnum.new("typeEmailReceipt")
+    HostedProfilePaymentOptions = SettingNameEnum.new("hostedProfilePaymentOptions")
   end
   
   # {AnetApi/xml/v1/schema/AnetApiSchema.xsd}ARBGetSubscriptionListSearchTypeEnum
@@ -3141,19 +3227,27 @@ end
   #   merchantAuthentication - MerchantAuthenticationType
   #   refId - SOAP::SOAPString
   #   customerProfileId - (any)
+  #   merchantCustomerId - SOAP::SOAPString
+  #   email - SOAP::SOAPString
   #   unmaskExpirationDate - SOAP::SOAPBoolean
   class GetCustomerProfileRequest 
     include ROXML
     xml_accessor :merchantAuthentication
     xml_accessor :refId
     xml_accessor :customerProfileId
+	xml_accessor :merchantCustomerId
+	xml_accessor :email
     xml_accessor :unmaskExpirationDate
+    xml_accessor :includeIssuerInfo
   
-    def initialize(merchantAuthentication = nil, refId = nil, customerProfileId = nil, unmaskExpirationDate = nil)
+    def initialize(merchantAuthentication = nil, refId = nil, customerProfileId = nil, unmaskExpirationDate = nil, merchantCustomerId = nil, email = nil, includeIssuerInfo = nil)
       @merchantAuthentication = merchantAuthentication
       @refId = refId
       @customerProfileId = customerProfileId
+	  @merchantCustomerId = merchantCustomerId
+	  @email = email
       @unmaskExpirationDate = unmaskExpirationDate
+      @includeIssuerInfo = includeIssuerInfo
     end
   end
   
@@ -3193,13 +3287,15 @@ end
     xml_accessor :customerProfileId
     xml_accessor :customerPaymentProfileId
     xml_accessor :unmaskExpirationDate
+    xml_accessor :includeIssuerInfo
   
-    def initialize(merchantAuthentication = nil, refId = nil, customerProfileId = nil, customerPaymentProfileId = nil, unmaskExpirationDate = nil)
+    def initialize(merchantAuthentication = nil, refId = nil, customerProfileId = nil, customerPaymentProfileId = nil, unmaskExpirationDate = nil, includeIssuerInfo = nil)
       @merchantAuthentication = merchantAuthentication
       @refId = refId
       @customerProfileId = customerProfileId
       @customerPaymentProfileId = customerPaymentProfileId
       @unmaskExpirationDate = unmaskExpirationDate
+      @includeIssuerInfo = includeIssuerInfo
     end
   end
   
@@ -3677,18 +3773,24 @@ end
   #   messages - MessagesType
   #   sessionToken - SOAP::SOAPString
   #   transaction - TransactionDetailsType
+  #   clientId - SOAP::SOAPString
+  #   transrefId - SOAP::SOAPString
   class GetTransactionDetailsResponse 
     include ROXML
     xml_accessor :refId
     xml_accessor :messages, :as => MessagesType
     xml_accessor :sessionToken
     xml_accessor :transaction, :as => TransactionDetailsType
+	xml_accessor :clientId
+	xml_accessor :transrefId
   
-    def initialize(refId = nil, messages = nil, sessionToken = nil, transaction = nil)
+    def initialize(refId = nil, messages = nil, sessionToken = nil, transaction = nil, clientId = nil, transrefId = nil)
       @refId = refId
       @messages = messages
       @sessionToken = sessionToken
       @transaction = transaction
+	  @clientId = clientId
+	  @transrefId = transrefId
     end
   end
   
@@ -3839,7 +3941,7 @@ end
   #   merchantAuthentication - MerchantAuthenticationType
   #   refId - SOAP::SOAPString
   #   customerProfileId - (any)
-  #   hostedProfileSettings - ArrayOfSetting
+  #   hostedProfileSettings - Settings
   class GetHostedProfilePageRequest 
     include ROXML
     xml_accessor :merchantAuthentication
@@ -3922,7 +4024,33 @@ end
       @totalNumInResultSet = totalNumInResultSet
     end
   end
+
+  # {AnetApi/xml/v1/schema/AnetApiSchema.xsd}getTransactionListForCustomerRequest
+  #   merchantAuthentication - MerchantAuthenticationType
+  #   refId - SOAP::SOAPString
+  #   customerProfileId - SOAP::SOAPString
+  #   customerPaymentProfileId - SOAP::SOAPString
+  #	  sorting - TransactionListSorting
+  #   paging - Paging
+  class GetTransactionListForCustomerRequest 
+    include ROXML
+    xml_accessor :merchantAuthentication
+    xml_accessor :refId
+	xml_accessor :customerProfileId
+	xml_accessor :customerPaymentProfileId
+	xml_accessor :sorting, :as => TransactionListSorting
+    xml_accessor :paging, :as => Paging
   
+    def initialize(merchantAuthentication = nil, refId = nil, customerProfileId = nil, customerPaymentProfileId = nil, sorting = nil, paging = nil)
+      @merchantAuthentication = merchantAuthentication
+      @refId = refId
+	  @customerProfileId = customerProfileId
+	  @customerPaymentProfileId = customerPaymentProfileId
+	  @sorting = sorting
+	  @paging = paging
+    end
+  end
+
   # {AnetApi/xml/v1/schema/AnetApiSchema.xsd}mobileDeviceRegistrationRequest
   #   merchantAuthentication - MerchantAuthenticationType
   #   refId - SOAP::SOAPString
@@ -4225,7 +4353,7 @@ end
   #   customerIP - SOAP::SOAPString
   #   cardholderAuthentication - CcAuthenticationType
   #   retail - TransRetailInfoType
-  #   transactionSettings - ArrayOfSetting
+  #   transactionSettings - Settings
   #   userFields - TransactionRequestType::UserFields
   class TransactionRequestType
     include ROXML 
@@ -4254,8 +4382,12 @@ end
     xml_accessor :retail, :as => TransRetailInfoType
     xml_accessor :transactionSettings, :as => Settings
     xml_accessor :userFields, :as => UserFields
+    xml_accessor :surcharge, :as => ExtendedAmountType
+    xml_accessor :merchantDescriptor
+    xml_accessor :subMerchant, :as => SubMerchantType
+    xml_accessor :tip, :as => ExtendedAmountType
   
-    def initialize(transactionType = nil, amount = nil, currencyCode = nil, payment = nil, profile = nil, solution = nil, callId = nil, authCode = nil, refTransId = nil, splitTenderId = nil, order = nil, lineItems = nil, tax = nil, duty = nil, shipping = nil, taxExempt = nil, poNumber = nil, customer = nil, billTo = nil, shipTo = nil, customerIP = nil, cardholderAuthentication = nil, retail = nil, transactionSettings = nil, userFields = nil)
+    def initialize(transactionType = nil, amount = nil, currencyCode = nil, payment = nil, profile = nil, solution = nil, callId = nil, authCode = nil, refTransId = nil, splitTenderId = nil, order = nil, lineItems = nil, tax = nil, duty = nil, shipping = nil, taxExempt = nil, poNumber = nil, customer = nil, billTo = nil, shipTo = nil, customerIP = nil, cardholderAuthentication = nil, retail = nil, transactionSettings = nil, userFields = nil, surcharge = nil, merchantDescriptor = nil, subMerchant = nil, tip = nil)
       @transactionType = transactionType
       @amount = amount
       @currencyCode = currencyCode
@@ -4281,6 +4413,10 @@ end
       @retail = retail
       @transactionSettings = transactionSettings
       @userFields = userFields
+      @surcharge = surcharge
+      @merchantDescriptor = merchantDescriptor
+      @subMerchant = subMerchant
+      @tip = tip
     end
   end
   
@@ -4473,8 +4609,9 @@ end
     xml_accessor :status
     xml_accessor :profile, :as => SubscriptionCustomerProfileType
     xml_accessor :order, :as => OrderType
+    xml_accessor :arbTransactions, :as => ARBTransactionList
   
-    def initialize(name = nil, paymentSchedule = nil, amount = nil, trialAmount = nil, status = nil, profile = nil, order = nil)
+    def initialize(name = nil, paymentSchedule = nil, amount = nil, trialAmount = nil, status = nil, profile = nil, order = nil, arbTransactions = nil)
       @name = name
       @paymentSchedule = paymentSchedule
       @amount = amount
@@ -4482,6 +4619,7 @@ end
       @status = status
       @profile = profile
       @order = order
+      @arbTransactions = arbTransactions
     end
   end
   
@@ -4489,16 +4627,19 @@ end
   #   merchantAuthentication - MerchantAuthenticationType
   #   refId - SOAP::SOAPString
   #   subscriptionId - (any)
+  #   includeTransactions - (any)
   class ARBGetSubscriptionRequest 
     include ROXML
     xml_accessor :merchantAuthentication, :as => MerchantAuthenticationType
     xml_accessor :refId
     xml_accessor :subscriptionId
+    xml_accessor :includeTransactions
   
-    def initialize(merchantAuthentication = nil, refId = nil, subscriptionId = nil)
+    def initialize(merchantAuthentication = nil, refId = nil, subscriptionId = nil, includeTransactions = nil)
       @merchantAuthentication = merchantAuthentication
       @refId = refId
       @subscriptionId = subscriptionId
+      @includeTransactions = includeTransactions
     end
   end
   
@@ -4736,13 +4877,27 @@ end
     end
   end
 
+ # {AnetApi/xml/v1/schema/AnetApiSchema.xsd}ArrayOfCardType 
+   class ArrayOfCardType < ::Array
+    include ROXML
+    xml_accessor :cardType
+
+    def initialize(cardType = [])
+     @cardType = cardType
+    end
+  end
+
   # {AnetApi/xml/v1/schema/AnetApiSchema.xsd}ProcessorType
   class ProcessorType
     include ROXML
     xml_accessor :name
+    xml_accessor :id
+    xml_accessor :cardTypes, :as => [ArrayOfCardType]
   
-    def initialize(name = nil)
+    def initialize(name = nil, id = nil, cardTypes = nil)
       @name = name
+      @id = id
+      @cardTypes = cardTypes
     end
   end
 
@@ -4810,6 +4965,8 @@ end
      @currency = currency
     end
   end
+
+  
   
   # {AnetApi/xml/v1/schema/AnetApiSchema.xsd}getMerchantDetailsResponse
   class GetMerchantDetailsResponse 
@@ -4825,8 +4982,9 @@ end
     xml_accessor :productCodes, :as => ArrayOfProductCode
     xml_accessor :paymentMethods, :as => ArrayOfPaymentMethod
     xml_accessor :currencies, :as => ArrayOfCurrencyCode
+    xml_accessor :publicClientKey
   
-    def initialize(refId = nil, messages = nil, sessionToken = nil, isTestMode = nil, processors = nil, merchantName = nil, gatewayId = nil, marketTypes = nil, productCodes = nil, paymentMethods = nil, currencies = nil)
+    def initialize(refId = nil, messages = nil, sessionToken = nil, isTestMode = nil, processors = nil, merchantName = nil, gatewayId = nil, marketTypes = nil, productCodes = nil, paymentMethods = nil, currencies = nil, publicClientKey = nil)
       @refId = refId
       @messages = messages
       @sessionToken = sessionToken
@@ -4838,6 +4996,35 @@ end
       @productCodes = productCodes
       @paymentMethods = paymentMethods
       @currencies = currencies
+      @publicClientKey = publicClientKey
+    end
+  end
+
+  # {AnetApi/xml/v1/schema/AnetApiSchema.xsd}updateMerchantDetailsRequest
+  class UpdateMerchantDetailsRequest
+    include ROXML
+    xml_accessor :merchantAuthentication
+    xml_accessor :refId
+    xml_accessor :isTestMode
+
+    def initialize(merchantAuthentication = nil, refId = nil, isTestMode = nil)
+      @merchantAuthentication = merchantAuthentication
+      @refId = refId
+      @isTestMode = isTestMode
+    end
+  end
+
+  # {AnetApi/xml/v1/schema/AnetApiSchema.xsd}updateMerchantDetailsResponse
+  class UpdateMerchantDetailsResponse
+    include ROXML
+    xml_accessor :merchantAuthentication, :as => MerchantAuthenticationType
+    xml_accessor :refId
+    xml_accessor :messages, :as => MessagesType
+
+    def initialize(merchantAuthentication = nil, refId = nil, messages = nil)
+      @merchantAuthentication = merchantAuthentication
+      @refId = refId
+      @messages = messages
     end
   end
 
@@ -4930,4 +5117,5 @@ end
       @token = token
     end
   end
+  
 end
